+++ packages/sync-server/app/createApp.js
@@ -28,7 +28,6 @@ export function createApp(ctx) {
   app.use(getLoggingMiddleware());
 
   app.use((req, res, next) => {
-    res.setHeader('X-Runtime', 'Tamanu Sync Server'); // TODO: deprecated
     res.setHeader('X-Tamanu-Server', 'Tamanu Sync Server');
     res.setHeader('X-Version', version);
     next();