+++ packages/sync-server/app/middleware/versionCompatibility.js
@@ -25,8 +25,7 @@ export const SUPPORTED_CLIENT_VERSIONS = {
 };
 
 export const versionCompatibility = (req, res, next) => {
-  // TODO: X-Runtime is deprecated
-  const clientType = req.header('X-Tamanu-Client') || req.header('X-Runtime');
+  const clientType = req.header('X-Tamanu-Client');
 
   if (!clientType) {
     // a thirdparty tool (or internal test suite) is using the API; ignore version checking